% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absent_tippR.R
\name{absent_tippr}
\alias{absent_tippr}
\title{Add tips to tree via user input}
\usage{
absent_tippr(tree, absent_list)
}
\arguments{
\item{tree}{Starting tree; object of type phylo}

\item{absent_list}{Vector of taxa in the total dataset that are not on the tree}
}
\value{
tree Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips not on existing tree to the tree via user input,
if they do not have congeners on the tree. This function will generate a
 tree, with nodes numbered. For each tip to be added, the user will be asked #'  to enter the number of the node they would like the tip to subtend.
}
\examples{
\dontrun{ new <- absent_tippr(tree, absent_list) }
}
