% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bears}
\alias{bears}
\title{bears}
\format{A collection of objects representing function inputs:
\describe{
    \item{tax_frame}{a two-column dataframe of taxon names and ages}
    \item{absent_list}{A list of taxa in the total data set, but are not on the tree}
    \item{mrca_df}{a two-column dataframe of taxa to be placed, and a set of taxa that indicate the MRCA which they will subtend}
    \item{tree}{A phylogenetic tree to which tips will be added}
}}
\source{
\url{https://github.com/revbayes/revbayes_tutorial/tree/master/RB_TotalEvidenceDating_FBD_Tutorial/data}
}
\usage{
data(bears)
}
\description{
A dataset containing 18 taxa and 62 characters for bears
}
\keyword{datasets}
