% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_loadR.R
\name{dataf_parsR}
\alias{dataf_parsR}
\title{Detect file format of taxon list (molecular, morphological, and stratigraphic) and maximum age of fossil, see RevBayes
total-evidence fossil file for an example of this.}
\usage{
dataf_parsR(dataf)
}
\arguments{
\item{dataf}{A data frame with one column containing the taxon name of tips in
phylogenetic tree, and one column indicating the maximum age of the tip. If
the tip is extant, use 0.0 as the age. Can be CSV or TSV.}
}
\value{
tax_list Dataframe containing the total set of tips on the tree
}
\description{
Detect file format of taxon list (molecular, morphological, and stratigraphic) and maximum age of fossil, see RevBayes
total-evidence fossil file for an example of this.
}
