% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fileFormat.R
\name{fileFormat}
\alias{fileFormat}
\title{Test file format is csv or tsv}
\usage{
fileFormat(dataf)
}
\arguments{
\item{dataf}{A data frame with one column containing the taxon name of tips in
phylogenetic tree, and one column indicating the maximum age of the tip. If
the tip is extant, use 0.0 as the age. Can be CSV or TSV.}
}
\value{
None
}
\description{
Test file format is csv or tsv
}
