% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genera_strippR.R
\name{genera_strippr}
\alias{genera_strippr}
\title{Check if tip is on tree}
\usage{
genera_strippr(tree, tax_list)
}
\arguments{
\item{tree}{Starting phylogeny, of type phylo}

\item{tax_list}{Total set of taxa on tree, as dataframe}
}
\value{
absent_list of taxa that are present in the total set of trees, but not the starting tree
}
\description{
Test which taxa in the total set of taxa are present on the
tree, and which need to be added.
}
\examples{
absent_taxa <- genera_strippr(tree, tax_list = tax_list)
}
