% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_absent_df.R
\name{make_absentdf}
\alias{make_absentdf}
\title{Determine which tips are not on the tree, and get their genera.}
\usage{
make_absentdf(absent_list)
}
\arguments{
\item{absent_list}{Vector of taxa in the total dataset that are not on the tree}
}
\value{
dataframe Dataframe objects expressing which tips are in the total set
        but not the tree, and their genera
}
\description{
Determine which tips are not on the tree, and get their genera.
}
