% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_df.R
\name{make_treedf}
\alias{make_treedf}
\title{Determine which tips are on the tree, and get their genera.}
\usage{
make_treedf(tree)
}
\arguments{
\item{tree}{Starting tree; object of type phylo}
}
\value{
dataframe Dataframe objects expressing the tips, and their genera
}
\description{
Determine which tips are on the tree, and get their genera.
}
