% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_molecular.R
\name{parse_moleculaR}
\alias{parse_moleculaR}
\title{Load molecular matrices to get taxon lists of each}
\usage{
parse_moleculaR(molDat)
}
\arguments{
\item{molDat}{A Nexus file containing molecular matrix}
}
\value{
mol_ns A vector of names present in either the matrix
}
\description{
Load molecular matrices to get taxon lists of each
}
