% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_morphological.R
\name{parse_morphology}
\alias{parse_morphology}
\title{Load morpohological matrices to get taxon lists}
\usage{
parse_morphology(morphDat)
}
\arguments{
\item{morphDat}{A Nexus file containing morphological matrix}
}
\value{
A vector of names present in either the matrix
}
\description{
Load morpohological matrices to get taxon lists
}
