% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_tippr.R
\name{present_tippr}
\alias{present_tippr}
\title{Add tips to the tree based on taxonomy}
\usage{
present_tippr(tree, absent_list)
}
\arguments{
\item{tree}{Starting tree; object of type phylo}

\item{absent_list}{Vector of taxa in the total dataset that are not on the tree}
}
\value{
tree. Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips not on existing tree to the tree via an MRCA node,
if they have congeners
}
\examples{
genera_tree <- present_tippr(tree, absent_list)
}
