% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_testr.R
\name{taxon_testr}
\alias{taxon_testr}
\title{Utility function to check taxon_names in parsed data file}
\usage{
taxon_testr(tax_frame)
}
\arguments{
\item{tax_frame}{Dataframe of taxa parsed from datafile.}
}
\value{
None
}
\description{
Utility function to check taxon_names in parsed data file
}
