% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dataf.R
\name{test_dataf}
\alias{test_dataf}
\title{Test if dataframe has two columns and a non-zero number of rows.}
\usage{
test_dataf(tax_list)
}
\arguments{
\item{tax_list}{Dataframe containing a list of taxa in the analysis and their ages.}
}
\value{
None
}
\description{
Test if dataframe has two columns and a non-zero number of rows.
}
