% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tippR.R
\name{tippR}
\alias{tippR}
\title{Add tips not on existing tree to the tree}
\usage{
tippR(absent, tree)
}
\arguments{
\item{absent}{Vector of taxa in the total dataset that are not on the tree}

\item{tree}{Starting tree; object of type phylo}
}
\value{
final_tree. Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips not on existing tree to the tree
}
