% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestartr.R
\docType{package}
\name{treestartr}
\alias{treestartr}
\alias{treestartr-package}
\title{treestartr: A package for generating reasonable starting trees for phylogenetic analysis.}
\description{
treestartr provides multiple ways to add tipts to a starting tree
}
\section{treestartr functions}{

absent_tippr, echo_rb, echo_subtree, present_tippr, rand_absent_tippr, text_placr
}

