% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absent_tippR.R
\name{absent_tippr}
\alias{absent_tippr}
\title{Add tips not on existing tree to the tree, if thye do not have congeners on the tree}
\usage{
absent_tippr(tree, absent_list)
}
\arguments{
\item{tree}{Starting tree; object of type phylo}

\item{absent_list}{Vector of taxa in the total dataset that are not on the tree}
}
\value{
tree. Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips not on existing tree to the tree, if thye do not have congeners on the tree
}
