% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fileFormat.R
\name{fileformat}
\alias{fileformat}
\title{Utility function to check if file is a csv or tsv.}
\usage{
fileformat(dataf)
}
\arguments{
\item{dataf}{Path to input file}
}
\value{
None
}
\description{
Utility function to check if file is a csv or tsv.
}
