% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genera_strippR.R
\name{genera_strippr}
\alias{genera_strippr}
\title{See which morphological & stratigraphic genera are represented
in the molecular tree}
\usage{
genera_strippr(tree, tax_list)
}
\arguments{
\item{tree}{Starting phylogeny, of type phylo}

\item{tax_list}{Total set of taxa on tree, as dataframe}
}
\value{
absent list of taxa that are present in the total set of trees, but not the
       starting tree
}
\description{
See which morphological & stratigraphic genera are represented
in the molecular tree
}
