% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus_parsR.R
\name{parse_morphology}
\alias{parse_morphology}
\title{Load morpohological and molecular matrices to get taxon lists of each}
\usage{
parse_morphology(morphDat)
}
\arguments{
\item{morphDat}{A Nexus file containing morphological matrix}

\item{molDat}{A Nexus file containing molecular matrix}
}
\value{
A vector of names present in either the matrix
}
\description{
Load morpohological and molecular matrices to get taxon lists of each
}
