% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_tippr.R
\name{present_tippR}
\alias{present_tippR}
\title{Add tips not on existing tree to the tree, if they have congeners}
\usage{
present_tippR(absent_list, tree)
}
\arguments{
\item{absent_list}{Vector of taxa in the total dataset that are not on the tree}

\item{tree}{Starting tree; object of type phylo}
}
\value{
tree. Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips not on existing tree to the tree, if they have congeners
}
