% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_absent_tippR.R
\name{rand_absent_tippR}
\alias{rand_absent_tippR}
\title{Add tips not on existing tree to the tree at random, if they do not have congeners on the tree}
\usage{
rand_absent_tippR(absent_list, tree)
}
\arguments{
\item{absent_list}{Vector of taxa in the total dataset that are not on the tree}

\item{tree}{Starting tree; object of type phylo}
}
\value{
tree. Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips not on existing tree to the tree at random, if they do not have congeners on the tree
}
