% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_placR.R
\name{text_placR}
\alias{text_placR}
\title{Add tips according to csv or tsv file of taxon names and taxa that form the clade
into which you'd like to insert the tip}
\usage{
text_placR(tree, mrca_df)
}
\arguments{
\item{tree}{Starting tree; object of type phylo}

\item{mrca_df}{Dataframe containing a column of the taxa you'd like to place,
and one column with the clade into which you'd like to place it}
}
\value{
tree. Phylo object containing the starting tree,
         and all tips that were added.
}
\description{
Add tips according to csv or tsv file of taxon names and taxa that form the clade
into which you'd like to insert the tip
}
